# Resolved conflicts:
#	ChangeLog
#	MSVC/build/3D.vcxproj
#	MSVC/build/bzadmin.vcxproj
#	MSVC/build/bzflag.vcxproj
#	MSVC/build/bzfs.vcxproj
#	MSVC/build/common.vcxproj
#	MSVC/build/date.vcxproj
#	MSVC/build/game.vcxproj
#	MSVC/build/geometry.vcxproj
#	MSVC/build/installer.vcxproj
#	MSVC/build/makehtml.vcxproj
#	MSVC/build/man2html.vcxproj
#	MSVC/build/mediafile.vcxproj
#	MSVC/build/net.vcxproj
#	MSVC/build/obstacle.vcxproj
#	MSVC/build/ogl.vcxproj
#	MSVC/build/platform.vcxproj
#	MSVC/build/scene.vcxproj
#	MSVC/bzadmin.sln
#	MSVC/bzflag.rc
#	MSVC/bzflag.sln
#	MSVC/bzfs.sln
#	MSVC/fullbuild.sln
#	README
#	README.WINDOWS
#	Xcode/BZFlag.xcodeproj/project.pbxproj
#	configure.ac
#	include/NetHandler.h
#	include/vectors.h
#	include/win32.h
#	package/win32/nsis/BZFlag.nsi
#	plugins/CustomZoneSample/CustomZoneSample.vcxproj
#	plugins/HoldTheFlag/HoldTheFlag.vcxproj
#	plugins/Phoenix/Phoenix.vcxproj
#	plugins/RogueGenocide/RogueGenocide.vcxproj
#	plugins/TimeLimit/TimeLimit.vcxproj
#	plugins/airspawn/airspawn.vcxproj
#	plugins/autoFlagReset/autoFlagReset.vcxproj
#	plugins/bzfscron/bzfscron.vcxproj
#	plugins/chathistory/chathistory.vcxproj
#	plugins/customflagsample/customflagsample.vcxproj
#	plugins/fairCTF/fairCTF.vcxproj
#	plugins/fastmap/fastmap.vcxproj
#	plugins/flagStay/flagStay.vcxproj
#	plugins/keepaway/keepaway.vcxproj
#	plugins/killall/killall.vcxproj
#	plugins/koth/koth.vcxproj
#	plugins/logDetail/logDetail.vcxproj
#	plugins/nagware/nagware.vcxproj
#	plugins/playHistoryTracker/playHistoryTracker.vcxproj
#	plugins/plugin_utils/plugin_utils.vcxproj
#	plugins/pushstats/pushstats.vcxproj
#	plugins/rabbitTimer/rabbitTimer.vcxproj
#	plugins/rabidRabbit/rabidRabbit.vcxproj
#	plugins/recordmatch/recordmatch.vcxproj
#	plugins/regFlag/regFlag.vcxproj
#	plugins/serverControl/serverControl.vcxproj
#	plugins/shockwaveDeath/shockwaveDeath.vcxproj
#	plugins/teamflagreset/teamflagreset.vcxproj
#	plugins/timedctf/timedctf.vcxproj
#	plugins/wwzones/wwzones.vcxproj
#	src/bzflag/LocalPlayer.cxx
#	src/bzflag/daylight.cxx
#	src/bzflag/effectsRenderer.cxx
#	src/bzflag/playing.cxx
#	src/date/buildDate.cxx

+++ src/bzfs/RecordReplay.cxx
@@ -1,5 +1,5 @@
 /* bzflag
- * Copyright (c) 1993-2016 Tim Riker
+ * Copyright (c) 1993-2017 Tim Riker
  *
  * This package is free software;  you can redistribute it and/or
  * modify it under the terms of the license found in the file
@@ -1666,9 +1666,9 @@ static void packVars(const std::string& key, void *data)
     pvd.len = sizeof(u16);
   }
 
-  pvd.buf = nboPackUByte(pvd.buf, key.length());
+  pvd.buf = nboPackUByte(pvd.buf, (uint8_t)key.length());
   pvd.buf = nboPackString(pvd.buf, key.c_str(), key.length());
-  pvd.buf = nboPackUByte(pvd.buf, value.length());
+  pvd.buf = nboPackUByte(pvd.buf, (uint8_t)value.length());
   pvd.buf = nboPackString(pvd.buf, value.c_str(), value.length());
   pvd.len += pairLen;
   pvd.count++;
