+++ src/game/CacheManager.cxx
@@ -1,5 +1,5 @@
 /* bzflag
- * Copyright (c) 1993-2016 Tim Riker
+ * Copyright (c) 1993-2017 Tim Riker
  *
  * This package is free software;  you can redistribute it and/or
  * modify it under the terms of the license found in the file
@@ -224,7 +224,7 @@ bool CacheManager::saveIndex()
 
   for (unsigned int i = 0; i < records.size(); i++) {
     const CacheRecord& rec = records[i];
-    fprintf(file, "%s\n%u %lu %lu ", rec.url.c_str(), rec.size, rec.date, rec.usedDate);
+    fprintf(file, "%s\n%u %llu %llu ", rec.url.c_str(), rec.size, (long long unsigned)rec.date, (long long unsigned)rec.usedDate);
     fprintf(file, "%s\n\n", rec.key.c_str());
   }
 
@@ -358,7 +358,7 @@ static bool compareUsedDate(const CacheManager::CacheRecord& a,
 
 /*
  * Local Variables: ***
- * mode:C ***
+ * mode: C ***
  * tab-width: 8 ***
  * c-basic-offset: 2 ***
  * indent-tabs-mode: t ***
